\name{e_dfclass}
\alias{e_dfclass}
\title{e_dfclass}
\usage{
e_dfclass()
}
\description{
Prints 'Explaining data ---'.
}
\examples{
e_dfclass()
df0: 자료1.csv
df1: 학업탈진
df2: GDP성장률과 지방세 성장률
df3: 감정노동자 서베이 데이터
df4: GDP, 소비, Lag소비
df5: 타이타닉
df6: 2018년 가계금융복지
df7: 카페 서베이
df8: 2017년 사업체 패널조사
df9: 국내여행 서베이
df10: 2022년 주거실태조사
df11: 2021년 NGO기부금자료
}
